# server/docs/swagger.yaml
definitions:
  data.BaseResponse:
    properties:
      code:
        type: integer
      data: {}
      err_msg:
        type: string
    type: object
  types.CreateReviewRequest:
    properties:
      content:
        type: string
      is_anonymous:
        type: boolean
      parentID:
        type: string
      pic_info:
        items:
          type: string
        type: array
      productID:
        type: integer
      stars:
        type: integer
    type: object
  types.LikeRequest:
    properties:
      review_id:
        type: string
    type: object
  types.ListReviewRequest:
    properties:
      product_id:
        type: integer
      stars:
        description: 0 means any stars
        type: integer
    type: object
  types.ListReviewResponse:
    properties:
      pinned_review:
        $ref: '#/definitions/types.ReviewInfo'
      review_list:
        items:
          $ref: '#/definitions/types.ReviewInfo'
        type: array
    type: object
  types.PinReviewRequest:
    properties:
      is_pinned:
        type: boolean
    type: object
  types.ReviewInfo:
    properties:
      content:
        type: string
      created_at:
        type: string
      current_user_liked:
        type: boolean
      id:
        type: string
      is_anonymous:
        type: boolean
      is_pinned:
        type: boolean
      likes:
        type: integer
      parent_id:
        type: string
      pic_info:
        items:
          type: string
        type: array
      product_id:
        type: integer
      stars:
        type: integer
      user_id:
        type: integer
    type: object
  types.UpdateReviewStatusRequest:
    properties:
      auto_flag:
        description: optional
        type: string
      is_harmful:
        description: optional
        type: boolean
      is_mismatch:
        description: optional
        type: boolean
      review_id:
        description: required
        type: string
      status:
        description: 'required: pending / processing / approved / hidden / rejected'
        type: string
    type: object
info:
  contact: {}
paths:
  /comment-ms/v1/customer/reviews:
    post:
      consumes:
      - application/json
      description: Create an review record.
      parameters:
      - description: CreateReviewRequest
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/types.CreateReviewRequest'
      - description: Client identifier
        enum:
        - customer
        - merchant
        in: path
        name: client
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/types.CreateReviewRequest'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  type: string
              type: object
      summary: Review Create
      tags:
      - Review
  /comment-ms/v1/customer/reviews/{review_id}/like:
    post:
      consumes:
      - application/json
      description: Like a review by id
      parameters:
      - description: LikeRequest
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/types.LikeRequest'
      - description: Client identifier
        enum:
        - customer
        - merchant
        in: path
        name: client
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  type: string
              type: object
      summary: Like a review
      tags:
      - Review
  /comment-ms/v1/customer/reviews/product/{product_id}:
    get:
      consumes:
      - application/json
      description: Get list of reviews for a product
      parameters:
      - description: Product ID
        in: path
        name: product_id
        required: true
        type: integer
      - description: Client identifier
        enum:
        - customer
        - merchant
        in: path
        name: client
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/types.ListReviewResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  type: string
              type: object
      summary: Get reviews by product
      tags:
      - Review
  /comment-ms/v1/customer/reviews/user:
    get:
      consumes:
      - application/json
      description: Get list of reviews created by current authenticated user
      parameters:
      - description: Client identifier
        enum:
        - customer
        - merchant
        in: path
        name: client
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/types.ReviewInfo'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  type: string
              type: object
      summary: Get reviews by user
      tags:
      - Review
  /comment-ms/v1/merchant/reviews/{review_id}:
    delete:
      consumes:
      - application/json
      description: Delete a review by id
      parameters:
      - description: Review ID
        in: path
        name: review_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete a review
      tags:
      - Review
    patch:
      consumes:
      - application/json
      description: Pin a review by id
      parameters:
      - description: PinReviewRequest
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/types.PinReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  type: string
              type: object
      summary: Pin a review
      tags:
      - Review
  /comment-ms/v1/merchant/reviews/{review_id}/replies:
    post:
      consumes:
      - application/json
      description: Reply an review record.
      parameters:
      - description: CreateReviewRequest
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/types.CreateReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/types.CreateReviewRequest'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  type: string
              type: object
      summary: Review Reply
      tags:
      - Review
  /comment-ms/v1/merchant/reviews/list:
    post:
      consumes:
      - application/json
      description: Filter reviews by product_id and stars (0 means any), ordered by
        created_at desc
      parameters:
      - description: ListReviewRequest
        in: body
        name: filter
        required: true
        schema:
          $ref: '#/definitions/types.ListReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/types.ReviewInfo'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  type: string
              type: object
      summary: List reviews by product and stars
      tags:
      - Review
  /comment-ms/v1/reviews/status:
    post:
      consumes:
      - application/json
      description: 'Update review fields: status (required), is_mismatch/is_harmful/auto_flag
        (optional). For agent service use.'
      parameters:
      - description: UpdateReviewStatusRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.UpdateReviewStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Review
      tags:
      - Review
  /comment-ms/v1/users/{user_id}/reviews:
    get:
      description: Get all reviews (excluding rejected) for a given user, ordered
        by created_at desc
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/types.ReviewInfo'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  type: string
              type: object
      summary: Get reviews by user ID
      tags:
      - Review
  /comment-ms/v1/reviews/status/{status}:
    get:
      consumes:
      - application/json
      description: Get all reviews with specified status.
      parameters:
      - description: Review status
        enum:
        - pending
        - processing
        - approved
        - hidden
        - rejected
        in: path
        name: status
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/types.ReviewInfo'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  type: string
              type: object
      summary: Get reviews by status
      tags:
      - Review
swagger: "2.0"